Network.BitTorrent.Transmission

Specialized types

type BandwidthPriority = TorrentPriority

type TorrentHash = Text

type Tag = Int

type TorrentList = Maybe [TorrentHash]

type TagGenerator = Text -> Value -> Tag

type SpeedLimit = Int

type FileIndices = [Int]

type PeerLimit = Int

type QueuePosition = Int

type TrackerAnnounce = Text

type TrackerId = Int

type SeedIdleLimit = Int

Idle limit in minutes

type SeedRatioLimit = Double

Ratio of upload:download. Between 2.0 and 5.0 recommended.

type Cookies = [(Text, Text)]

A collection och key-value pairs for use as cookies for torrents

type SeedIdleMode = SettingsMode

type SeedRatioMode = SettingsMode

type TorrentFileId = Int

data TorrentFile

A representation of a file. Use fileLength, fileName, and fileCompleted.

data TorrentFileStats

Another representation of a file. Use fileWanted, filePriority, and fileCompleted.

data TorrentFileIds

A collection of TorrentFileIds.

data TorrentFiles

A collection of TorrentFiles.

data DownloadDirectory

A directory path where downloads are to be placed.

data TorrentDate

Yeah, I don't really know what this is. Might be epoch in ms (or s)? TODO: CHECK THIS OUT!

data TorrentField t where

A field of information about a torrent. The field is accompanied with the type of that field to make some compile-time guarantees by the type system.

Constructors

TorrentActivityDate :: TorrentField TorrentDate

TorrentAddedDate :: TorrentField TorrentDate

TorrentBandwidthPriority :: TorrentField BandwidthPriority

TorrentComment :: TorrentField Text

TorrentCorruptEver :: TorrentField Int

TorrentCreator :: TorrentField Text

TorrentDateCreated :: TorrentField TorrentDate

TorrentDesiredAvailable :: TorrentField Int

TorrentDoneDate :: TorrentField TorrentDate

TorrentDownloadDir :: TorrentField Text

TorrentDownloadedEver :: TorrentField Int

TorrentDownloadLimit :: TorrentField SpeedLimit

TorrentDownloadLimited :: TorrentField Bool

TorrentError :: TorrentField Int

TorrentErrorString :: TorrentField Text

TorrentETA :: TorrentField Int

TorrentETAIdle :: TorrentField Int

TorrentFiles :: TorrentField TorrentFiles

TorrentFileStats :: TorrentField TorrentStatsList

TorrentFilesWanted :: TorrentField TorrentFileIds

TorrentFilesUnwanted :: TorrentField TorrentFileIds

TorrentHashString :: TorrentField Text

TorrentHaveUnchecked :: TorrentField Int

TorrentHaveValid :: TorrentField Int

TorrentHonorsSessionLimits :: TorrentField Bool

TorrentId :: TorrentField Int

TorrentIsFinished :: TorrentField Bool

TorrentIsPrivate :: TorrentField Bool

TorrentIsStalled :: TorrentField Bool

TorrentLeftUntilDone :: TorrentField Int

TorrentLocation :: TorrentField Text

TorrentMagnetLink :: TorrentField Text

TorrentManualAnnounceTime :: TorrentField Int

TorrentMaxConnectedPeers :: TorrentField Int

TorrentMetadataPercentComplete :: TorrentField Double

TorrentName :: TorrentField Text

TorrentPeerLimit :: TorrentField Int

TorrentPeersConnected :: TorrentField Int

TorrentPeersGettingFromUs :: TorrentField Int

TorrentPeersSendingToUs :: TorrentField Int

TorrentPercentDone :: TorrentField Double

TorrentPieceCount :: TorrentField Int

TorrentPieceSize :: TorrentField Int

TorrentPriorities :: TorrentField [TorrentFilePriority]

TorrentPriorityHigh :: TorrentField TorrentFileIds

TorrentPriorityNormal :: TorrentField TorrentFileIds

TorrentPriorityLow :: TorrentField TorrentFileIds

TorrentQueuePosition :: TorrentField Int

TorrentRateDownload :: TorrentField Int

TorrentRateUpload :: TorrentField Int

TorrentRecheckProgress :: TorrentField Double

TorrentSecondsDownloading :: TorrentField Int

TorrentSecondsSeeding :: TorrentField Int

TorrentSeedIdleLimit :: TorrentField SeedIdleLimit

TorrentSeedIdleMode :: TorrentField SettingsMode

TorrentSeedRatioLimit :: TorrentField SeedRatioLimit

TorrentSeedRatioMode :: TorrentField SettingsMode

TorrentSizeWhenDone :: TorrentField Int64

TorrentStartDate :: TorrentField TorrentDate

TorrentStatus :: TorrentField TorrentStatus

TorrentTrackerAdd :: TorrentField Text

TorrentTrackerRemove :: TorrentField TrackerId

TorrentTrackers :: TorrentField [TorrentTracker]

TorrentTrackerStats :: TorrentField [TorrentTrackerStats]

TorrentTotalSize :: TorrentField Int64

TorrentFile :: TorrentField Text

TorrentUploadedEver :: TorrentField Int64

TorrentUploadLimit :: TorrentField Int

TorrentUploadLimited :: TorrentField Bool

TorrentUploadRatio :: TorrentField Double

TorrentWanted :: TorrentField [Bool]

TorrentWebseeds :: TorrentField [Text]

TorrentWebseedsSendingToUs :: TorrentField Int

data TorrentStatus

The current status of a torrent.

Constructors

TorrentStatusCheckWait

TorrentStatusChecking

TorrentStatusDownloading

TorrentStatusSeeding

TorrentStatusFinished

data TorrentFields

A number of TorrentFields with associated values.

data TorrentFieldsRequest

A number of TorrentFields to be requested from Transmission.

data TorrentInfoList

A TorrentInfoList is a number of (>= 0) torrents represented by TorrentFields. These may be extracted using torrentInfoToList.

data Request t where

A request to be sent to Transmission. Includes the type of the expected return value.

Constructors

TorrentStart :: TorrentList -> Request ()

TorrentStop :: TorrentList -> Request ()

TorrentVerify :: TorrentList -> Request ()

TorrentSet :: TorrentList -> TorrentFields -> Request ()

TorrentGet :: TorrentList -> TorrentFieldsRequest -> Request TorrentInfoList

TorrentAdd :: Text -> DownloadDirectory -> Maybe Cookies -> Request (Either TorrentDuplicate TorrentAdded)

data TorrentAdded

Response when a torrent has been successfully added. Use torrentFields in combination with lookupTorrentField to get more info about the torrent.

data TorrentDuplicate

Response when a torrent attempted to be added but it already existed. Use torrentFields in combination with lookupTorrentField to get more info about the torrent.

data TransmissionConnection

A connection to Transmission. Contains the state. Use transmissionConnect to obtain.

data Response t

A response to a sent Request. Might result in a responseError or contains a responseValue. The type of the responseValue, if any, is strictly typed as to be the same as the requested type. The typesystem should be able to do much work for us.

class TorrentListable t a | t -> a where

Methods

asList :: t -> [a]

type classes

class FileCompleted a where

General typeclass for a representation for a file where a number of bytes of completion is appropriate.

Methods

fileCompleted :: a -> Int64

How many bytes of the file has been downloaded. When used with TorrentFile the result may be compared with fileLength to see if file is fully completed.

data SettingsMode

A mode where the settings for the torrent should be fetched from.

class (Eq t, Typeable t, FromJSON t, ToJSON t) => Responsable t

class TorrentFieldsContainer t where

Any container which contains TorrentFields.

Methods

torrentFields :: t -> TorrentFields

torrentFieldsToList :: t -> [(Text, Value)]

stuff...

allTorrentFileIds :: TorrentFileIds

A list of TorrentFileIds representing all files.

fileLength :: TorrentFile -> Int64

The length (in bytes) of a file.

fileName :: TorrentFile -> Text

The name of a file.

fileWanted :: TorrentFileStats -> Bool

Whether or not this file should be downloaded.

filePriority :: TorrentFileStats -> Int

The download priority of a file.

unspecifiedDirectory :: DownloadDirectory

A standard DownloadDirectory for used when no specific directory should be used. In the current version this is a directory unsorted in the default directory for downloads.

downloadTo :: [String] -> DownloadDirectory

A list of directories where the downloads are to be placed. If not absolute this is relative to the default downloads directory specified by Transmission.

(+|+) :: Responsable t => TorrentFieldsRequest -> TorrentField t -> TorrentFieldsRequest

Adds a torrent field to a request.

addRequest :: Responsable t => TorrentFieldsRequest -> TorrentField t -> TorrentFieldsRequest

Alias for '(+|+)'.

makeFieldsRequest :: Responsable t => TorrentField t -> TorrentFieldsRequest

Create a request with a single field.

allTorrentFields :: TorrentFieldsRequest

A request asking for all fields possible.

lookupTorrentField :: forall t. Responsable t => TorrentField t -> TorrentFields -> Maybe t

Find the content of a specific field.

torrentInfoToList :: TorrentInfoList -> [TorrentFields]

Extract the TorrentInfoList to a list of TorrentFields.

allTorrents :: TorrentList

A TorrentList representing all torrents.

setTorrentField :: Responsable t => TorrentField t -> t -> TorrentFields -> TorrentFields

Sets a specific field in a collection of torrent fields.

torrentFieldName :: TorrentField t -> Text

The textual name for a field as used in communication with Transmission.

transmissionConnect :: URI -> IO TransmissionConnection

Connect to a Transmission daemon running at a known URL.

Example:

 con <- transmissionConnect $ parseURI "http://localhost:9091/transmission/rpc"

sendRequest :: Responsable t => TransmissionConnection -> TagGenerator -> Request t -> IO (Response t)

Send a Request and await Response.

responseError :: Response t -> Maybe Text

responseTag :: Response t -> Maybe Tag

responseValue :: Response t -> Maybe t