Network.BitTorrent.Transmission.Internal

Documentation

type TorrentPriority = Int

type BandwidthPriority = TorrentPriority

type TorrentFilePriority = TorrentPriority

type TorrentHash = Text

type Tag = Int

type TorrentList = Maybe [TorrentHash]

type TagGenerator = Text -> Value -> Tag

type SpeedLimit = Int

type FileIndices = [Int]

type PeerLimit = Int

type QueuePosition = Int

type TrackerAnnounce = Text

type TrackerId = Int

type SeedIdleLimit = Int

Idle limit in minutes

type SeedRatioLimit = Double

Ratio of upload:download. Between 2.0 and 5.0 recommended.

type Cookies = [(Text, Text)]

A collection och key-value pairs for use as cookies for torrents

type SeedIdleMode = SettingsMode

type SeedRatioMode = SettingsMode

type M = HashMap

type TorrentFileId = Int

newtype TorrentFile

A representation of a file. Use fileLength, fileName, and fileCompleted.

Constructors

TorrentFileData (Int64, Int64, Text)

newtype TorrentFileStats

Another representation of a file. Use fileWanted, filePriority, and fileCompleted.

Constructors

TorrentFileStts (Int64, Bool, Int)

newtype TorrentFileIds

A collection of TorrentFileIds.

Constructors

TorrentFileIds (Maybe [TorrentFileId])

newtype TorrentTracker

Simple information about a tracker.

Constructors

Tracker (Text, Int, Text, Int)

data TorrentTrackerStats

In-depth information about a tracker. For basic use TorrentTracker is the better option.

Constructors

TrackerStats

Fields

trackerStatsAnnounce :: Text

announceState :: Int

trackerDownloadCount :: Int

hasAnnounced :: Bool

hasScraped :: Bool

trackerHost :: Text

trackerStatsId :: TrackerId

trackerIsBackup :: Bool

lastAnnouncePeerCount :: Int

lastAnnounceResult :: Text

lastAnnounceStartTime :: TorrentDate

lastAnnounceSucceeded :: Bool

lastAnnounceTime :: Int

lastAnnounceTimedOut :: Bool

lastScrapeResult :: Text

lastScrapeStartTime :: TorrentDate

lastScrapeSucceeded :: Bool

lastScrapeTime :: Int

lastScrapeTimedOut :: Bool

trackerLeecherCount :: Int

nextAnnounceTime :: TorrentDate

nextScrapeTime :: TorrentDate

trackerStatsScrape :: Text

scrapeState :: Int

trackerSeederCount :: Int

trackerStatsTier :: Int

class TorrentTrackerData tracker where

Methods

announce :: tracker -> Text

trackerId :: tracker -> TrackerId

scrape :: tracker -> Text

trackerTier :: tracker -> Int

newtype TorrentStatsList

A collection of TorrentFiles.

Constructors

TorrentStatsList [TorrentFileStats]

class TorrentListable t a | t -> a where

Methods

asList :: t -> [a]

newtype TorrentFiles

A collection of TorrentFiles.

Constructors

TorrentFileCol [TorrentFile]

allTorrentFileIds :: TorrentFileIds

A list of TorrentFileIds representing all files.

fileLength :: TorrentFile -> Int64

The length (in bytes) of a file.

fileName :: TorrentFile -> Text

The name of a file.

fileWanted :: TorrentFileStats -> Bool

Whether or not this file should be downloaded.

filePriority :: TorrentFileStats -> Int

The download priority of a file.

class FileCompleted a where

General typeclass for a representation for a file where a number of bytes of completion is appropriate.

Methods

fileCompleted :: a -> Int64

How many bytes of the file has been downloaded. When used with TorrentFile the result may be compared with fileLength to see if file is fully completed.

data SettingsMode

A mode where the settings for the torrent should be fetched from.

Constructors

ModeDisabled

ModeTorrent

ModeSession

unspecifiedDirectory :: DownloadDirectory

A standard DownloadDirectory for used when no specific directory should be used. In the current version this is a directory unsorted in the default directory for downloads.

newtype DownloadDirectory

A directory path where downloads are to be placed.

Constructors

DownloadDirectory [String]

downloadTo :: [String] -> DownloadDirectory

A list of directories where the downloads are to be placed. If not absolute this is relative to the default downloads directory specified by Transmission.

newtype TorrentDate

Yeah, I don't really know what this is. Might be epoch in ms (or s)? TODO: CHECK THIS OUT!

Constructors

TorrentDate Int

torrentDate :: TorrentDate -> UTCTime

data TorrentField t where

A field of information about a torrent. The field is accompanied with the type of that field to make some compile-time guarantees by the type system.

Constructors

TorrentActivityDate :: TorrentField TorrentDate

TorrentAddedDate :: TorrentField TorrentDate

TorrentBandwidthPriority :: TorrentField BandwidthPriority

TorrentComment :: TorrentField Text

TorrentCorruptEver :: TorrentField Int

TorrentCreator :: TorrentField Text

TorrentDateCreated :: TorrentField TorrentDate

TorrentDesiredAvailable :: TorrentField Int

TorrentDoneDate :: TorrentField TorrentDate

TorrentDownloadDir :: TorrentField Text

TorrentDownloadedEver :: TorrentField Int

TorrentDownloadLimit :: TorrentField SpeedLimit

TorrentDownloadLimited :: TorrentField Bool

TorrentError :: TorrentField Int

TorrentErrorString :: TorrentField Text

TorrentETA :: TorrentField Int

TorrentETAIdle :: TorrentField Int

TorrentFiles :: TorrentField TorrentFiles

TorrentFileStats :: TorrentField TorrentStatsList

TorrentFilesWanted :: TorrentField TorrentFileIds

TorrentFilesUnwanted :: TorrentField TorrentFileIds

TorrentHashString :: TorrentField Text

TorrentHaveUnchecked :: TorrentField Int

TorrentHaveValid :: TorrentField Int

TorrentHonorsSessionLimits :: TorrentField Bool

TorrentId :: TorrentField Int

TorrentIsFinished :: TorrentField Bool

TorrentIsPrivate :: TorrentField Bool

TorrentIsStalled :: TorrentField Bool

TorrentLeftUntilDone :: TorrentField Int

TorrentLocation :: TorrentField Text

TorrentMagnetLink :: TorrentField Text

TorrentManualAnnounceTime :: TorrentField Int

TorrentMaxConnectedPeers :: TorrentField Int

TorrentMetadataPercentComplete :: TorrentField Double

TorrentName :: TorrentField Text

TorrentPeerLimit :: TorrentField Int

TorrentPeersConnected :: TorrentField Int

TorrentPeersGettingFromUs :: TorrentField Int

TorrentPeersSendingToUs :: TorrentField Int

TorrentPercentDone :: TorrentField Double

TorrentPieceCount :: TorrentField Int

TorrentPieceSize :: TorrentField Int

TorrentPriorities :: TorrentField [TorrentFilePriority]

TorrentPriorityHigh :: TorrentField TorrentFileIds

TorrentPriorityNormal :: TorrentField TorrentFileIds

TorrentPriorityLow :: TorrentField TorrentFileIds

TorrentQueuePosition :: TorrentField Int

TorrentRateDownload :: TorrentField Int

TorrentRateUpload :: TorrentField Int

TorrentRecheckProgress :: TorrentField Double

TorrentSecondsDownloading :: TorrentField Int

TorrentSecondsSeeding :: TorrentField Int

TorrentSeedIdleLimit :: TorrentField SeedIdleLimit

TorrentSeedIdleMode :: TorrentField SettingsMode

TorrentSeedRatioLimit :: TorrentField SeedRatioLimit

TorrentSeedRatioMode :: TorrentField SettingsMode

TorrentSizeWhenDone :: TorrentField Int64

TorrentStartDate :: TorrentField TorrentDate

TorrentStatus :: TorrentField TorrentStatus

TorrentTrackerAdd :: TorrentField Text

TorrentTrackerRemove :: TorrentField TrackerId

TorrentTrackers :: TorrentField [TorrentTracker]

TorrentTrackerStats :: TorrentField [TorrentTrackerStats]

TorrentTotalSize :: TorrentField Int64

TorrentFile :: TorrentField Text

TorrentUploadedEver :: TorrentField Int64

TorrentUploadLimit :: TorrentField Int

TorrentUploadLimited :: TorrentField Bool

TorrentUploadRatio :: TorrentField Double

TorrentWanted :: TorrentField [Bool]

TorrentWebseeds :: TorrentField [Text]

TorrentWebseedsSendingToUs :: TorrentField Int

data TorrentStatus

The current status of a torrent.

Constructors

TorrentStatusCheckWait

TorrentStatusChecking

TorrentStatusDownloading

TorrentStatusSeeding

TorrentStatusFinished

data AnyTorrentField

Constructors

forall t . (Typeable t, Responsable t) => AnyTorrentField (TorrentField t)

anyTorrentField :: AnyTorrentField -> forall t. (FromJSON t, Typeable t, Responsable t) => TorrentField t

anyTorrentField' :: forall t. (FromJSON t, Typeable t, Responsable t) => AnyTorrentField -> Maybe (TorrentField t)

mapAllTorrentFields :: (forall t. TorrentField t -> b) -> [b]

newtype TorrentFields

A number of TorrentFields with associated values.

Constructors

TorrentFields (M Text Value)

newtype TorrentFieldsRequest

A number of TorrentFields to be requested from Transmission.

Constructors

TorrentFieldsRequest [Text]

(+|+) :: Responsable t => TorrentFieldsRequest -> TorrentField t -> TorrentFieldsRequest

Adds a torrent field to a request.

addRequest :: Responsable t => TorrentFieldsRequest -> TorrentField t -> TorrentFieldsRequest

Alias for '(+|+)'.

makeFieldsRequest :: Responsable t => TorrentField t -> TorrentFieldsRequest

Create a request with a single field.

fieldSettable :: TorrentField t -> Bool

Test if a TorrentField is able to be set. (That is; it may be used with TorrentSet.)

fieldGettable :: TorrentField t -> Bool

Test if a TorrentField is able to get. (That is; it may be used with TorrentGet.)

allTorrentFields :: TorrentFieldsRequest

A request asking for all fields possible.

lookupTorrentField :: forall t. Responsable t => TorrentField t -> TorrentFields -> Maybe t

Find the content of a specific field.

newtype TorrentInfoList

A TorrentInfoList is a number of (>= 0) torrents represented by TorrentFields. These may be extracted using torrentInfoToList.

Constructors

TorrentInfoList [TorrentFields]

torrentInfoToList :: TorrentInfoList -> [TorrentFields]

Extract the TorrentInfoList to a list of TorrentFields.

data Request t where

A request to be sent to Transmission. Includes the type of the expected return value.

Constructors

TorrentStart :: TorrentList -> Request ()

TorrentStop :: TorrentList -> Request ()

TorrentVerify :: TorrentList -> Request ()

TorrentSet :: TorrentList -> TorrentFields -> Request ()

TorrentGet :: TorrentList -> TorrentFieldsRequest -> Request TorrentInfoList

TorrentAdd :: Text -> DownloadDirectory -> Maybe Cookies -> Request (Either TorrentDuplicate TorrentAdded)

data RawRequest

A general request as to be sent to Transmission.

Constructors

RawReq

Fields

rawMethod :: Text

rawArguments :: Value

rawTag :: Tag

data RawResponse

A general response as received from Transmission.

Constructors

RawResp

Fields

rawError :: Maybe Text

rawResponseTag :: Maybe Tag

responseObject :: Maybe Value

allTorrents :: TorrentList

A TorrentList representing all torrents.

class TorrentFieldsContainer t where

Any container which contains TorrentFields.

Methods

torrentFields :: t -> TorrentFields

torrentFieldsToList :: t -> [(Text, Value)]

newtype TorrentAdded

Response when a torrent has been successfully added. Use torrentFields in combination with lookupTorrentField to get more info about the torrent.

Constructors

TorrentAdded TorrentFields

newtype TorrentDuplicate

Response when a torrent attempted to be added but it already existed. Use torrentFields in combination with lookupTorrentField to get more info about the torrent.

Constructors

TorrentDuplicate TorrentFields

setTorrentField :: Responsable t => TorrentField t -> t -> TorrentFields -> TorrentFields

Sets a specific field in a collection of torrent fields.

torrentFieldName :: TorrentField t -> Text

The textual name for a field as used in communication with Transmission.

rawRequest :: Responsable t => TagGenerator -> Request t -> RawRequest

Transforms an abstract Request to a more general RawRequest which may be serialized to JSON and sent in communication with Transmission.

rawRequest' :: TagGenerator -> Text -> Value -> RawRequest

Create a general RawRequest from a request name and an arguments object.

joinCookies :: Cookies -> Text

Utility function to join a series of cookies to a single text string.

type TransmissionId = Text

A session id used for communication with Transmission.

newtype TransmissionConnection

A connection to Transmission. Contains the state. Use transmissionConnect to obtain.

Constructors

TCon (URI, MVar (ByteString -> Request ByteString, BrowserState (HandleStream ByteString)))

transmissionConnect :: URI -> IO TransmissionConnection

Connect to a Transmission daemon running at a known URL.

Example:

 con <- transmissionConnect $ parseURI "http://localhost:9091/transmission/rpc"

rqf :: URI -> String -> ByteString -> Request ByteString

A requestfunction used for simplicity when updating a TransmissionConnection with a valid session. This specifies how to actually send the JSON request.

runBrowserActionUnsafe :: TransmissionConnection -> BrowserAction (HandleStream ByteString) r -> IO r

Run some special session or browser action within an active connection. Since this may be used to do anything to affect the connection it is named with the Unsafe postfix.

sessionHeader :: HeaderName

The HeaderName used to contain the TransmissionId.

sendRawRequest :: TransmissionConnection -> RawRequest -> IO RawResponse

Sends a general RawRequest using an active TransmissionConnection and awaits its response.

sendRequest :: Responsable t => TransmissionConnection -> TagGenerator -> Request t -> IO (Response t)

Send a Request and await Response.

data Response t

A response to a sent Request. Might result in a responseError or contains a responseValue. The type of the responseValue, if any, is strictly typed as to be the same as the requested type. The typesystem should be able to do much work for us.

Constructors

Response

Fields

responseError :: Maybe Text

responseTag :: Maybe Tag

responseValue :: Maybe t

class (Eq t, Typeable t, ResponseType t flag) => Responsable' flag t where

Helper class to make the typesystem strict but still be able to handle some non trivial JSON responses.

Methods

rawResponse'' :: flag -> Value -> Parser t

responseJSONValue' :: flag -> t -> Value

class (Eq t, Typeable t, FromJSON t, ToJSON t) => Responsable t where

Methods

rawResponse :: RawResponse -> Response t

rawResponse' :: Value -> Parser t

responseJSONValue :: t -> Value

data KeyedResponse

Helper datatype used for creating the possibility of non trivial JSON responses and still have a strict response type.

data SimpleResponse

Helper datatype used for creating the possibility of non trivial JSON responses and still have a strict response type.

data NoResponse

Helper datatype used for creating the possibility of non trivial JSON responses and still have a strict response type.

data EitherResponse

Helper datatype used for creating the possibility of non trivial JSON responses and still have a strict response type.

class (Eq t, Typeable t, FromJSON t, ToJSON t) => ResponseType t flag | t -> flag where

Helper class to make the typesystem strict but still be able to handle some non trivial JSON responses.

Methods

responseTypeValue :: flag -> t -> Text

responseTypeUndefined :: t -> flag

magnetToHash :: Text -> Maybe TorrentHash

Extract the hash from a magnet link.